refactor: [M3-6377] - MUI v5 - Components > ProductInformationBanner
