+++ packages/manager/src/components/ProductInformationBanner/ProductInformationBanner.tsx
@@ -11,42 +11,42 @@ interface Props {
   bannerLocation: ProductInformationBannerLocation;
 }
 
-type CombinedProps = Props & Partial<NoticeProps>;
+type ProductInformationBannerProps = Props & Partial<NoticeProps>;
 
-const ProductInformationBanner = (props: CombinedProps) => {
-  const { bannerLocation, ...rest } = props;
-  const { productInformationBanners } = useFlags();
+export const ProductInformationBanner = React.memo(
+  (props: ProductInformationBannerProps) => {
+    const { bannerLocation, ...rest } = props;
+    const { productInformationBanners } = useFlags();
 
-  const thisBanner = (productInformationBanners ?? []).find(
-    (thisBanner) => thisBanner.bannerLocation === bannerLocation
-  );
-
-  if (!thisBanner) {
-    return null;
-  }
-
-  let hasBannerExpired = true;
-  try {
-    hasBannerExpired = isAfter(
-      new Date().toISOString(),
-      thisBanner?.expirationDate
+    const thisBanner = (productInformationBanners ?? []).find(
+      (thisBanner) => thisBanner.bannerLocation === bannerLocation
     );
-  } catch (err) {
-    reportException(err);
+
+    if (!thisBanner) {
+      return null;
+    }
+
+    let hasBannerExpired = true;
+    try {
+      hasBannerExpired = isAfter(
+        new Date().toISOString(),
+        thisBanner?.expirationDate
+      );
+    } catch (err) {
+      reportException(err);
+    }
+
+    if (hasBannerExpired) {
+      return null;
+    }
+
+    return (
+      <DismissibleBanner
+        preferenceKey={`${bannerLocation}-${thisBanner.expirationDate}`}
+        {...rest}
+      >
+        <HighlightedMarkdown textOrMarkdown={thisBanner.message} />
+      </DismissibleBanner>
+    );
   }
-
-  if (hasBannerExpired) {
-    return null;
-  }
-
-  return (
-    <DismissibleBanner
-      preferenceKey={`${bannerLocation}-${thisBanner.expirationDate}`}
-      {...rest}
-    >
-      <HighlightedMarkdown textOrMarkdown={thisBanner.message} />
-    </DismissibleBanner>
-  );
-};
-
-export default React.memo(ProductInformationBanner);
+);
