--- packages/manager/src/components/ProductInformationBanner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProductInformationBanner';
