+++ packages/manager/src/features/Databases/DatabaseLanding/DatabaseLanding.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
 import Hidden from 'src/components/core/Hidden';
 import LandingHeader from 'src/components/LandingHeader';
-import ProductInformationBanner from 'src/components/ProductInformationBanner';
+import { ProductInformationBanner } from 'src/components/ProductInformationBanner/ProductInformationBanner';
 import { CircleProgress } from 'src/components/CircleProgress';
 import { DatabaseEmptyState } from './DatabaseEmptyState';
 import { DatabaseInstance } from '@linode/api-v4/lib/databases';
